chore: move things around. need to fix types still
