+++ packages/react-query/src/shared/proxy/decorationProxy.ts
@@ -1,7 +1,7 @@
 import { AnyRouter } from '@trpc/server';
 import { createRecursiveProxy } from '@trpc/server/shared';
 import { getArrayQueryKey } from '../../internals/getArrayQueryKey';
-import { getQueryKey } from '../../internals/getQueryKey';
+import { getQueryKeyInternal } from '../../internals/getQueryKey';
 import { CreateReactQueryHooks } from '../hooks/createHooksInternal';
 
 /**
@@ -28,7 +28,7 @@ export function createReactProxyDecoration<
     }
     const [input, ...rest] = args;
 
-    const queryKey = getQueryKey(path, input);
+    const queryKey = getQueryKeyInternal(path, input);
 
     // Expose queryKey helper
     if (lastArg === 'getQueryKey') {
